split up multer config from middleware
