+++ package.json
@@ -21,6 +21,7 @@
         "@types/express": "^4.17.17",
         "@types/jest": "^29.5.14",
         "@types/jsonwebtoken": "^9.0.9",
+        "@types/multer": "^1.4.12",
         "@types/node": "^22.13.13",
         "@types/supertest": "^6.0.3",
         "@types/swagger-ui-express": "^4.1.8",
